Ignore tasks which had been created before profiling.
Besides, if the task was created but hadn't been started yet, the start time and
the end time will be assigned to the end time of profiling. If the task was
created and was still running at the end of profiling, it's end time will be
assigned to the end time of profiling.